# docs/api-contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/groups/{groupId}/diary/context": {
      "get": {
        "description": "Returns the group's active custom questions and the current user's diary entry (if any) for the specified date. The `date` parameter must be the caller's **local date** in `YYYY-MM-DD` format so that users in different timezones each see their own correct diary day.",
        "operationId": "DiaryController_getDiaryContext",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": true,
            "in": "query",
            "description": "The diary date in the user's local timezone (YYYY-MM-DD).",
            "schema": {
              "format": "date",
              "example": "2026-08-26",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Active group questions and the current user's diary entry (if any) for the given date.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiaryContextResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "date must be a valid ISO 8601 date string (YYYY-MM-DD)."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "Group membership required."
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Get diary context for a given date",
        "tags": [
          "Diary"
        ]
      }
    },
    "/groups/{groupId}/diary/feed": {
      "get": {
        "description": "Returns every group member paired with their diary entry (and answers) for the specified date. Members who have not written an entry for that date are included with `entry: null`. Requires group membership.",
        "operationId": "DiaryController_getGroupDailyFeed",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": true,
            "in": "query",
            "description": "The diary date in the user's local timezone (YYYY-MM-DD).",
            "schema": {
              "format": "date",
              "example": "2026-08-26",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Group members' diary entries for the given date.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupDailyFeedResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "date must be a valid ISO 8601 date string (YYYY-MM-DD)."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "Group membership required."
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Get all group members' diary entries for a given date",
        "tags": [
          "Diary"
        ]
      }
    },
    "/groups/{groupId}/diary/answers": {
      "post": {
        "description": "Creates the caller's answer for a custom question on the given date. If the user has no diary entry for that date yet, one is created automatically.",
        "operationId": "DiaryController_createAnswer",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAnswerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Diary answer created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswerResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Request payload is invalid or an unsupported question type was provided."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "Group membership required."
          },
          "404": {
            "description": "Group question not found or does not belong to this group."
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Create a diary answer for a given date",
        "tags": [
          "Diary"
        ]
      }
    },
    "/groups/{groupId}/diary/answers/{answerId}": {
      "patch": {
        "operationId": "DiaryController_updateAnswer",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "answerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAnswerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Diary answer updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswerResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Request payload is invalid."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "Group membership required or diary ownership is violated."
          },
          "404": {
            "description": "Answer not found."
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Update an existing diary answer",
        "tags": [
          "Diary"
        ]
      }
    },
    "/groups/{groupId}/diary/entries/{diaryEntryId}/photos/upload-url": {
      "post": {
        "description": "Returns a short-lived S3 URL for directly uploading a photo to a diary entry. Upload the image using the requested MIME type before the URL expires.",
        "operationId": "DiaryController_requestDiaryPhotoUpload",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "diaryEntryId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestDiaryPhotoUploadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Presigned diary photo upload URL created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresignedUploadResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Request payload is invalid or the image type or size is unsupported."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "Group membership required or diary ownership is violated."
          },
          "404": {
            "description": "Diary entry not found."
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Request a presigned diary photo upload URL",
        "tags": [
          "Diary"
        ]
      }
    },
    "/groups/{groupId}/diary/entries/{diaryEntryId}/photos": {
      "post": {
        "description": "Registers metadata for a photo that was uploaded directly to S3 using a server-generated diary photo storage key.",
        "operationId": "DiaryController_registerDiaryPhoto",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "diaryEntryId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDiaryPhotoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Diary photo registered successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Request payload is invalid or the storage key does not belong to this diary entry."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "Group membership required or diary ownership is violated."
          },
          "404": {
            "description": "Diary entry not found."
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Register uploaded diary photo metadata",
        "tags": [
          "Diary"
        ]
      }
    },
    "/groups/{groupId}/diary/entries/{diaryEntryId}/photos/{photoId}": {
      "delete": {
        "operationId": "DiaryController_deleteDiaryPhoto",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "diaryEntryId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "photoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Diary photo deleted successfully."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "Group membership required or diary ownership is violated."
          },
          "404": {
            "description": "Photo not found."
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Delete a diary photo",
        "tags": [
          "Diary"
        ]
      }
    },
...
  },
  "info": {
    "title": "MARU API",
    "description": "MARU backend API documentation",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "maru_session"
      }
    },
    "schemas": {
...
      "GroupQuestionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "groupId": {
            "type": "string"
          },
          "question": {
            "type": "string"
          },
          "displayOrder": {
            "type": "number"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdByUserId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "groupId",
          "question",
          "displayOrder",
          "isActive",
          "createdByUserId",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "DailyQuestionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "question": {
            "type": "string"
          },
          "questionDate": {
            "type": "string",
            "description": "ISO-8601 date string (date only)",
            "example": "2026-08-26"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "question",
          "questionDate",
          "createdAt"
        ]
      },
      "AnswerResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "diaryEntryId": {
            "type": "string"
          },
          "questionType": {
            "type": "string",
            "enum": [
              "CUSTOM",
              "DAILY"
            ]
          },
          "groupQuestionId": {
            "type": "object",
            "nullable": true
          },
          "body": {
            "type": "string"
          },
          "questionSnapshot": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "diaryEntryId",
          "questionType",
          "body",
          "questionSnapshot",
          "createdAt",
          "updatedAt"
        ]
      },
      "PhotoResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "diaryEntryId": {
            "type": "string"
          },
          "uploadedByUserId": {
            "type": "string"
          },
          "storageKey": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          },
          "width": {
            "type": "number"
          },
          "height": {
            "type": "number"
          },
          "sizeBytes": {
            "type": "number"
          },
          "displayOrder": {
            "type": "number"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "diaryEntryId",
          "uploadedByUserId",
          "storageKey",
          "mimeType",
          "width",
          "height",
          "sizeBytes",
          "displayOrder",
          "createdAt"
        ]
      },
      "DiaryEntryContextDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "diaryDate": {
            "type": "string",
            "description": "ISO-8601 date string (date only)"
          },
          "answers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AnswerResponseDto"
            }
          },
          "photos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PhotoResponseDto"
            }
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "diaryDate",
          "answers",
          "photos",
          "createdAt",
          "updatedAt"
        ]
      },
      "DiaryContextResponseDto": {
        "type": "object",
        "properties": {
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GroupQuestionResponseDto"
            }
          },
          "dailyQuestion": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/DailyQuestionResponseDto"
              }
            ]
          },
          "entry": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/DiaryEntryContextDto"
              }
            ]
          }
        },
        "required": [
          "questions"
        ]
      },
      "FeedMemberUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "profileImageKey": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "FeedEntryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "diaryDate": {
            "type": "string",
            "description": "ISO-8601 date string (date only)"
          },
          "answers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AnswerResponseDto"
            }
          },
          "photos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PhotoResponseDto"
            }
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "diaryDate",
          "answers",
          "photos",
          "createdAt",
          "updatedAt"
        ]
      },
      "FeedMemberEntryDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/FeedMemberUserDto"
          },
          "entry": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/FeedEntryDto"
              }
            ]
          }
        },
        "required": [
          "userId",
          "user"
        ]
      },
      "GroupDailyFeedResponseDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "description": "ISO-8601 date string (date only)"
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FeedMemberEntryDto"
            }
          }
        },
        "required": [
          "date",
          "members"
        ]
      },
      "CreateAnswerDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "description": "The diary date in the user's local timezone (YYYY-MM-DD).",
            "example": "2026-08-26",
            "format": "date"
          },
          "questionType": {
            "type": "string",
            "enum": [
              "CUSTOM",
              "DAILY"
            ],
            "description": "Question type: CUSTOM for group questions, DAILY for the global daily question.",
            "example": "CUSTOM"
          },
          "groupQuestionId": {
            "type": "string",
            "description": "Group custom question id. Required for CUSTOM question type.",
            "nullable": true
          },
          "body": {
            "type": "string",
            "description": "Answer text",
            "maxLength": 2000,
            "example": "Today I enjoyed dinner with my family."
          }
        },
        "required": [
          "date",
          "questionType",
          "body"
        ]
      },
      "UpdateAnswerDto": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string",
            "description": "Updated answer text",
            "maxLength": 2000,
            "example": "Edited answer text."
          }
        },
        "required": [
          "body"
        ]
      },
      "RequestDiaryPhotoUploadDto": {
        "type": "object",
        "properties": {
          "mimeType": {
            "type": "string",
            "description": "Image MIME type.",
            "enum": [
              "image/jpeg",
              "image/png",
              "image/webp"
            ],
            "example": "image/jpeg"
          },
          "sizeBytes": {
            "type": "number",
            "description": "Image size in bytes.",
            "minimum": 1,
            "maximum": 10485760,
            "example": 1048576
          }
        },
        "required": [
          "mimeType",
          "sizeBytes"
        ]
      },
      "PresignedUploadResponseDto": {
        "type": "object",
        "properties": {
          "uploadUrl": {
            "type": "string",
            "description": "Short-lived S3 URL for directly uploading the image.",
            "format": "uri"
          },
          "storageKey": {
            "type": "string",
            "description": "Server-generated S3 object key for the uploaded image."
          }
        },
        "required": [
          "uploadUrl",
          "storageKey"
        ]
      },
      "RegisterDiaryPhotoDto": {
        "type": "object",
        "properties": {
          "storageKey": {
            "type": "string",
            "description": "Server-generated S3 object key returned by upload-url.",
            "example": "diary-entries/entry-id/photos/550e8400-e29b-41d4-a716-446655440000.jpg"
          },
          "mimeType": {
            "type": "string",
            "description": "Uploaded image MIME type.",
            "enum": [
              "image/jpeg",
              "image/png",
              "image/webp"
            ],
            "example": "image/jpeg"
          },
          "width": {
            "type": "number",
            "description": "Uploaded image width in pixels.",
            "minimum": 1,
            "example": 1200
          },
          "height": {
            "type": "number",
            "description": "Uploaded image height in pixels.",
            "minimum": 1,
            "example": 900
          },
          "sizeBytes": {
            "type": "number",
            "description": "Uploaded image size in bytes.",
            "minimum": 1,
            "maximum": 10485760,
            "example": 1048576
          }
        },
        "required": [
          "storageKey",
          "mimeType",
          "width",
          "height",
          "sizeBytes"
        ]
      }
    }
  }
}